Wrote an error message
I anticipate that failing to isolate the hard drive’s device ID is likely to be a common problem, and the mount command’s error message in that case probably won’t be very helpful. So we want to intercept that case earlier in the process and provide our own error instead.